build: fix sns dependencies on nns (#278)
# Motivation

Sns-js should not have dependencies on nns-js.

# Changes

- fix sns-js dependencies on nns-js
- move and expose `assertPercentageNumber` to `@dfinity/utils`
