+++ packages/sns/src/governance.canister.spec.ts
@@ -1,8 +1,9 @@
 import type { ActorSubclass } from "@dfinity/agent";
-import { Vote } from "@dfinity/nns";
-import { InvalidPercentageError } from "@dfinity/nns/src";
 import { Principal } from "@dfinity/principal";
-import { arrayOfNumberToUint8Array } from "@dfinity/utils";
+import {
+  arrayOfNumberToUint8Array,
+  InvalidPercentageError,
+} from "@dfinity/utils";
 import { mock } from "jest-mock-extended";
 import type {
   ListNervousSystemFunctionsResponse,
@@ -22,6 +23,7 @@ import {
   SnsNeuronPermissionType,
   SnsProposalDecisionStatus,
   SnsProposalRewardStatus,
+  SnsVote,
 } from "./enums/governance.enums";
 import { SnsGovernanceError } from "./errors/governance.errors";
 import { SnsGovernanceCanister } from "./governance.canister";
@@ -868,7 +870,7 @@ describe("Governance canister", () => {
     const proposalId = {
       id: 123n,
     };
-    const vote = Vote.Yes;
+    const vote = SnsVote.Yes;
     const params: SnsRegisterVoteParams = {
       neuronId: neuronIdMock,
       vote,
